Source directory not correctly passed as argument
The sphinx-build was not being passed the source directory correctly when the parameter SPHINXOPTS was left unset. In all honesty, I do not know why this small change has fixed the bug, but I've tested this and it now builds the documentation when I run

```
make html
```

from windows `cmd`.